requirements.txt
eodhd>=1.0.7
google-auth>=2.12.0
google-cloud-bigquery>=3.3.3
pandas>=1.3.5
requests>=2.28.1
streamlit>=1.13.0